// openpectus/frontend/src/app/api/model/processUnit.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { State } from './state';


/**
 * Represents a process unit. 
 */
export interface ProcessUnit { 
    id: number;
    name: string;
    state: State;
    location?: string;
    runtime_msec?: number;
}